pyproject.toml
[tool.poetry]
name = "api-nomades"
version = "0.1.0"
description = ""
authors = ["FelipePy <felipespsampaiomotasp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
bcrypt = "^4.1.2"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
fastapi = "^0.105.0"
email-validator = "^2.1.0.post1"
uvicorn = "^0.24.0.post1"
aiocron = "^1.8"
python-jose = "^3.3.0"
loguru = "^0.7.2"
requests = "^2.31.0"
pytest = "^7.4.4"
httpx = "^0.26.0"
beautifulsoup4 = "^4.12.3"
python-multipart = "^0.0.17"
taskipy = "^1.14.0"
blue = "^0.9.1"
isort = "^5.13.2"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
package-mode = false
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
lint = "blue app --check --diff . && isort app --check --diff ."
apply-lint = "blue app . && isort app ."

pre_run-dev = "docker compose -f infra/docker-compose.yml up -d pgbouncer"
run-dev = "uvicorn bootstrap.server:app --reload"

pre_run = "docker compose -f infra/docker-compose.yml up -d pgbouncer && alembic upgrade head"
run = "docker compose -f infra/docker-compose.yml up -d nomades"

pre_run-watch = "docker compose -f infra/docker-compose.yml up -d pgbouncer && alembic upgrade head"
run-watch = "docker compose -f infra/docker-compose.yml up nomades"

stop = "docker compose -f infra/docker-compose.yml down"

pre_test = "docker compose -f infra/docker-compose.yml up -d psql_test"
test = "pytest tests -s -x -vv --cov=app"
post_test = "coverage html && docker compose -f infra/docker-compose.yml stop psql_test"
